+++ extensions/mssql/src/features.ts
@@ -875,7 +875,7 @@ export class SqlAssessmentServicesFeature extends SqlOpsFeature<undefined> {
 	protected registerProvider(options: undefined): Disposable {
 		const client = this._client;
 
-		let assessmentInvoke = async (ownerUri: string, targetType: number): Promise<azdata.SqlAssessmentResult> => {
+		let assessmentInvoke = async (ownerUri: string, targetType: azdata.SqlAssessmentTargetType): Promise<azdata.SqlAssessmentResult> => {
 			let params: contracts.SqlAssessmentParams = { ownerUri: ownerUri, targetType: targetType };
 			try {
 				return client.sendRequest(contracts.SqlAssessmentInvokeRequest.type, params);
@@ -887,7 +887,7 @@ export class SqlAssessmentServicesFeature extends SqlOpsFeature<undefined> {
 			return undefined;
 		};
 
-		let getAssessmentItems = async (ownerUri: string, targetType: number): Promise<azdata.SqlAssessmentResult> => {
+		let getAssessmentItems = async (ownerUri: string, targetType: azdata.SqlAssessmentTargetType): Promise<azdata.SqlAssessmentResult> => {
 			let params: contracts.SqlAssessmentParams = { ownerUri: ownerUri, targetType: targetType };
 			try {
 				return client.sendRequest(contracts.GetSqlAssessmentItemsRequest.type, params);
