+++ src/sql/azdata.proposed.d.ts
@@ -86,6 +86,7 @@ declare module 'azdata' {
 
 	export namespace dataprotocol {
 		export function registerSerializationProvider(provider: SerializationProvider): vscode.Disposable;
+		export function registerSqlAssessmentServicesProvider(provider: SqlAssessmentServicesProvider): vscode.Disposable;
 	}
 
 	export interface HyperlinkComponent {
@@ -398,6 +399,44 @@ declare module 'azdata' {
 
 	export interface TaskInfo {
 		targetLocation?: string;
+	}
+	// Assessment interfaces
+	export enum SqlAssessmentTargetType {
+		Server = 1,
+		Database = 2
+	}
+
+	export enum SqlAssessmentResultItemKind {
+		RealResult = 0,
+		Warning = 1,
+		Error = 2
+	}
+
+	export interface SqlAssessmentResultItem {
+		rulesetVersion: string;
+		rulesetName: string;
+		targetType: SqlAssessmentTargetType;
+		targetName: string;
+		checkId: string;
+		tags: string[];
+		displayName: string;
+		description: string;
+		message: string;
+		helpLink: string;
+		level: string;
+		timestamp: string;
+		kind: SqlAssessmentResultItemKind;
+	}
+
+	export interface SqlAssessmentResult extends ResultStatus {
+		items: SqlAssessmentResultItem[];
+		apiVersion: string;
+	}
+
+	export interface SqlAssessmentServicesProvider extends DataProvider {
+		assessmentInvoke(ownerUri: string, targetType: SqlAssessmentTargetType): Promise<SqlAssessmentResult>;
+		getAssessmentItems(ownerUri: string, targetType: SqlAssessmentTargetType): Promise<SqlAssessmentResult>;
+		generateAssessmentScript(items: SqlAssessmentResultItem[]): Promise<ResultStatus>;
 	}
 }
